+++ design/components/app.component.ts
@@ -1,9 +1,4 @@
-import { Component, method, applyTemplate } from '@apexdesigner/dsl/component';
-import { SchemaFormsService } from '@apexdesigner/schema-forms';
-import { SelectUserFieldComponent } from '@components';
-import { SelectRoleFieldComponent } from '@components';
-import { SelectRoleNameFieldComponent } from '@components';
-import { SfReferenceFieldComponent } from '@components';
+import { Component, applyTemplate } from '@apexdesigner/dsl/component';
 import { AuthService, PackageService } from '@services';
 
 /**
@@ -17,18 +12,6 @@ export class AppComponent extends Component {
 
   /** Package Service */
   packageService!: PackageService;
-
-  /** Schema Forms Service */
-  schemaFormsService!: SchemaFormsService;
-
-  /** Register Fields - Register custom schema form fields */
-  @method({ callOnLoad: true })
-  registerFields(): void {
-    this.schemaFormsService.registerField('select-user', () => Promise.resolve(SelectUserFieldComponent as any));
-    this.schemaFormsService.registerField('select-role', () => Promise.resolve(SelectRoleFieldComponent as any));
-    this.schemaFormsService.registerField('select-role-name', () => Promise.resolve(SelectRoleNameFieldComponent as any));
-    this.schemaFormsService.registerField('reference', () => Promise.resolve(SfReferenceFieldComponent as any));
-  }
 }
 
 applyTemplate(AppComponent, [